+++ config.go
@@ -167,6 +167,8 @@ type config struct {
 
 	SoReuseport bool
 
+	PathPrefix string
+
 	MaxSrcDimension    int
 	MaxSrcResolution   int
 	MaxSrcFileSize     int
@@ -221,6 +223,10 @@ type config struct {
 	WatermarkPath    string
 	WatermarkURL     string
 	WatermarkOpacity float64
+
+	FallbackImageData string
+	FallbackImagePath string
+	FallbackImageURL  string
 
 	NewRelicAppName string
 	NewRelicKey     string
@@ -294,6 +300,8 @@ func configure() error {
 	boolEnvConfig(&conf.CacheControlPassthrough, "IMGPROXY_CACHE_CONTROL_PASSTHROUGH")
 
 	boolEnvConfig(&conf.SoReuseport, "IMGPROXY_SO_REUSEPORT")
+
+	strEnvConfig(&conf.PathPrefix, "IMGPROXY_PATH_PREFIX")
 
 	intEnvConfig(&conf.MaxSrcDimension, "IMGPROXY_MAX_SRC_DIMENSION")
 	megaIntEnvConfig(&conf.MaxSrcResolution, "IMGPROXY_MAX_SRC_RESOLUTION")
@@ -372,6 +380,10 @@ func configure() error {
 	strEnvConfig(&conf.WatermarkPath, "IMGPROXY_WATERMARK_PATH")
 	strEnvConfig(&conf.WatermarkURL, "IMGPROXY_WATERMARK_URL")
 	floatEnvConfig(&conf.WatermarkOpacity, "IMGPROXY_WATERMARK_OPACITY")
+
+	strEnvConfig(&conf.FallbackImageData, "IMGPROXY_FALLBACK_IMAGE_DATA")
+	strEnvConfig(&conf.FallbackImagePath, "IMGPROXY_FALLBACK_IMAGE_PATH")
+	strEnvConfig(&conf.FallbackImageURL, "IMGPROXY_FALLBACK_IMAGE_URL")
 
 	strEnvConfig(&conf.NewRelicAppName, "IMGPROXY_NEW_RELIC_APP_NAME")
 	strEnvConfig(&conf.NewRelicKey, "IMGPROXY_NEW_RELIC_KEY")