+++ vips.h
@@ -66,7 +66,9 @@ int vips_flip_horizontal_go(VipsImage *in, VipsImage **out);
 
 int vips_extract_area_go(VipsImage *in, VipsImage **out, int left, int top, int width, int height);
 int vips_smartcrop_go(VipsImage *in, VipsImage **out, int width, int height);
-int vips_trim(VipsImage *in, VipsImage **out, double threshold);
+int vips_trim(VipsImage *in, VipsImage **out, double threshold,
+              gboolean smart, double r, double g, double b,
+              gboolean equal_hor, gboolean equal_ver);
 
 int vips_gaussblur_go(VipsImage *in, VipsImage **out, double sigma);
 int vips_sharpen_go(VipsImage *in, VipsImage **out, double sigma);
@@ -87,7 +89,6 @@ int vips_pngsave_go(VipsImage *in, void **buf, size_t *len, int interlace, int q
 int vips_webpsave_go(VipsImage *in, void **buf, size_t *len, int quality, gboolean strip);
 int vips_gifsave_go(VipsImage *in, void **buf, size_t *len);
 int vips_icosave_go(VipsImage *in, void **buf, size_t *len);
-int vips_heifsave_go(VipsImage *in, void **buf, size_t *len, int quality);
 int vips_bmpsave_go(VipsImage *in, void **buf, size_t *len);
 int vips_tiffsave_go(VipsImage *in, void **buf, size_t *len, int quality);